requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altgraph==0.17.4
anyio==4.4.0
async-timeout==4.0.3     
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30       
cffi==1.17.1
charset-normalizer==3.3.2
discord.py==2.4.0        
exceptiongroup==1.2.2    
ffmpeg-python==0.2.0     
frozenlist==1.4.1        
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
multidict==6.1.0
mutagen==1.47.0
packaging==24.1
pefile==2024.8.26
pycparser==2.22
pycryptodomex==3.20.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyNaCl==1.5.0
pywin32-ctypes==0.2.3
requests==2.32.3
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.2.3
websockets==13.0.1
yarl==1.11.1
youtube-dl==2021.12.17
youtube-search-python==1.6.6
yt-dlp==2024.8.6
zipp==3.20.2